@@ -11,17 +11,17 @@
   "author": "Alex Adamant",
   "license": "ISC",
   "dependencies": {
-    "@prisma/client": "^6.19.0",
+    "@prisma/client": "^6.19.2",
     "@volley/db": "*",
     "express": "^5.1.0",
     "ioredis": "^5.8.2",
     "telegraf": "^4.15.3"
   },
   "devDependencies": {
-    "@types/express": "^5.0.5",
-    "@types/node": "^18.16.1",
-    "prettier": "^3.1.1",
+    "@types/express": "^5.0.6",
+    "@types/node": "^24.10.0",
+    "prettier": "^3.8.1",
     "ts-node": "^10.9.1",
-    "typescript": "^5.3.3"
+    "typescript": "^5.9.3"
   }
 }
