@@ -1,49 +1,417 @@
 <script lang="ts">
+  import { page } from "$app/state";
   import { goto } from "$app/navigation";
-  import { page } from "$app/stores";
+  import { resolve } from "$app/paths";
+  import type { Pathname } from "$app/types";
+  import { browser } from "$app/environment";
+  import { onMount } from "svelte";
+  import { SvelteURLSearchParams } from "svelte/reactivity";
   import { twMerge } from "tailwind-merge";
+  import PrimaryNav from "$lib/components/primary-nav.svelte";
+  import * as Select from "$lib/components/ui/select";
 
-  export let data;
+  let { data } = $props();
 
-  $: slug = $page.params.slug;
+  type StatusValue = "active" | "all";
+
+  const normalizeStatus = (value: string): StatusValue =>
+    value === "all" ? "all" : "active";
+
+  let rangeValue = $state("all");
+  let statusValue = $state<StatusValue>("active");
+
+  const slug = $derived(page.params.slug ?? "");
+  const rangeStorageKey = $derived(`volley-range:${slug}`);
+  const statusStorageKey = $derived(`volley-status:${slug}`);
+
+  const rangeOptions = $derived(
+    data.rangeOptions.map((range) => ({
+      value: range.key,
+      label: range.label,
+    })),
+  );
+  const rangeLabel = $derived(
+    rangeOptions.find((option) => option.value === rangeValue)?.label ??
+      "Range",
+  );
+  const statusOptions = [
+    { value: "active", label: "Active" },
+    { value: "all", label: "All" },
+  ];
+  const statusLabel =
+    statusOptions.find((option) => option.value === statusValue)?.label ??
+    "Status";
+
+  let lastRangeKey = $state("all");
+  let lastStatusValue = $state<StatusValue>("active");
+
+  $effect(() => {
+    const next = data.activeRange?.key ?? "all";
+    if (next !== lastRangeKey) {
+      lastRangeKey = next;
+      rangeValue = next;
+    }
+  });
+
+  $effect(() => {
+    const next = normalizeStatus(data.status ?? "active");
+    if (next !== lastStatusValue) {
+      lastStatusValue = next;
+      statusValue = next;
+    }
+  });
+
+  const toPathname = (value: string) => value as Pathname;
+
+  const buildQuery = (rangeKey: string, statusKey: string) => {
+    const params = new SvelteURLSearchParams(page.url.searchParams);
+    if (rangeKey) {
+      params.set("range", rangeKey);
+    } else {
+      params.delete("range");
+    }
+    if (statusKey) {
+      params.set("status", statusKey);
+    } else {
+      params.delete("status");
+    }
+    const queryString = params.toString();
+    return queryString ? `?${queryString}` : "";
+  };
+
+  const applyRange = (nextRange: string) => {
+    rangeValue = nextRange;
+    if (browser) {
+      localStorage.setItem(rangeStorageKey, nextRange);
+    }
+    goto(
+      resolve(
+        toPathname(`${page.url.pathname}${buildQuery(nextRange, statusValue)}`),
+      ),
+      {
+        keepFocus: true,
+        noScroll: true,
+      },
+    );
+  };
+
+  const applyStatus = (nextStatus: string) => {
+    const normalized = normalizeStatus(nextStatus);
+    statusValue = normalized;
+    if (browser) {
+      localStorage.setItem(statusStorageKey, normalized);
+    }
+    goto(
+      resolve(
+        toPathname(`${page.url.pathname}${buildQuery(rangeValue, normalized)}`),
+      ),
+      {
+        keepFocus: true,
+        noScroll: true,
+      },
+    );
+  };
+
+  const handleRangeChange = (nextRange: string) => {
+    applyRange(nextRange);
+  };
+
+  const handleStatusChange = (nextStatus: string) => {
+    applyStatus(nextStatus);
+  };
+
+  type MatchTooltip = {
+    x: number;
+    y: number;
+    detail: {
+      result: string;
+      score: string;
+      teammates: string[];
+      opponents: string[];
+    };
+  };
+
+  let matchTooltip = $state<MatchTooltip | null>(null);
+
+  const showMatchTooltip = (
+    event: MouseEvent | FocusEvent,
+    detail: MatchTooltip["detail"],
+  ) => {
+    const target = event.currentTarget as HTMLElement | null;
+    if (!target) return;
+    const rect = target.getBoundingClientRect();
+    const offset = 12;
+    const tooltipWidth = 192;
+    const tooltipHeight = 72;
+    const preferredX = rect.left - tooltipWidth - offset;
+    const fallbackX = rect.right + offset;
+    let x = preferredX;
+    let y = rect.top + rect.height / 2 - tooltipHeight / 2;
+
+    if (browser) {
+      const maxX = window.innerWidth - tooltipWidth - offset;
+      const maxY = window.innerHeight - tooltipHeight - offset;
+      x = preferredX < offset ? Math.min(fallbackX, maxX) : preferredX;
+      x = Math.min(Math.max(offset, x), maxX);
+      y = Math.min(Math.max(offset, y), maxY);
+    }
+
+    matchTooltip = {
+      x,
+      y,
+      detail,
+    };
+  };
+
+  const hideMatchTooltip = () => {
+    matchTooltip = null;
+  };
+
+  const navItems = $derived.by(() => {
+    const query = buildQuery(rangeValue, statusValue);
+    const playersPath = `/chat/${slug}`;
+    return [
+      {
+        label: "Players",
+        href: toPathname(`${playersPath}${query}`),
+        active: page.url.pathname === playersPath,
+      },
+      {
+        label: "Teams",
+        href: toPathname(`/chat/${slug}/team-stats${query}`),
+        active: page.url.pathname === `/chat/${slug}/team-stats`,
+      },
+      {
+        label: "League Stats",
+        href: toPathname(`/chat/${slug}/league-stats${query}`),
+        active: page.url.pathname === `/chat/${slug}/league-stats`,
+      },
+      {
+        label: "Results",
+        href: toPathname(`/chat/${slug}/day-results${query}`),
+        active: page.url.pathname === `/chat/${slug}/day-results`,
+      },
+      {
+        label: "Admin",
+        href: toPathname(`/chat/${slug}/admin${query}`),
+        active: page.url.pathname === `/chat/${slug}/admin`,
+      },
+    ];
+  });
+
+  onMount(() => {
+    if (!browser) return;
+    const params = new SvelteURLSearchParams(page.url.searchParams);
+    const storedRange = localStorage.getItem(rangeStorageKey);
+    const storedStatus = localStorage.getItem(statusStorageKey);
+    let changed = false;
+
+    if (!params.get("range") && storedRange) {
+      params.set("range", storedRange);
+      rangeValue = storedRange;
+      changed = true;
+    }
+
+    if (!params.get("status") && storedStatus) {
+      const normalized = normalizeStatus(storedStatus);
+      params.set("status", normalized);
+      statusValue = normalized;
+      changed = true;
+    }
+
+    if (changed) {
+      goto(resolve(toPathname(`${page.url.pathname}?${params.toString()}`)), {
+        replaceState: true,
+        keepFocus: true,
+        noScroll: true,
+      });
+    }
+  });
 </script>
 
-<div class="flex flex-row">
-  <div class="w-2/3">1</div>
-  <table class="w-1/3">
-    <thead>
-      <tr class="text-left [&>th]:pr-4">
-        <th>№</th>
-        <th>Name</th>
-        <th>Games</th>
-        <th>Rating</th>
-        <th>Change</th>
-      </tr>
-    </thead>
-    <tbody>
-      {#each data.results as item, index}
-        <tr
-          class="cursor-pointer hover:bg-gray-100"
-          on:click={() => {
-            goto(`/chat/${slug}/user/${item.id}`);
-          }}
+<section
+  class="border-stroke shadow-card sticky top-3 z-[200] rounded-2xl border bg-white/90 p-3 backdrop-blur"
+>
+  <div class="flex flex-wrap items-center gap-3">
+    <PrimaryNav items={navItems} />
+
+    <div class="ml-auto flex flex-wrap items-center gap-2">
+      <Select.Root
+        value={rangeValue}
+        onValueChange={handleRangeChange}
+        type="single"
+      >
+        <Select.Trigger
+          size="sm"
+          class="border-stroke text-ink h-8 w-44 rounded-full bg-white/80 px-3 text-xs font-semibold"
         >
-          <td>{index + 1}</td>
-          <td>{item.name}</td>
-          <td>{item.games}</td>
-          <td>{item.rating}</td>
-          <td
-            class={twMerge(
-              "pr-7 text-right",
-              item.ratingChange > 0 && "text-green-500",
-              item.ratingChange < 0 && "text-red-500",
-            )}
-            >{item.ratingChange > 0
-              ? `+${item.ratingChange}`
-              : item.ratingChange}</td
-          >
+          <span class="truncate">{rangeLabel}</span>
+        </Select.Trigger>
+        <Select.Content class="border-stroke bg-white">
+          {#each rangeOptions as option (option.value)}
+            <Select.Item value={option.value} label={option.label} />
+          {/each}
+        </Select.Content>
+      </Select.Root>
+
+      <Select.Root
+        value={statusValue}
+        onValueChange={handleStatusChange}
+        type="single"
+      >
+        <Select.Trigger
+          size="sm"
+          class="border-stroke text-ink h-8 w-28 rounded-full bg-white/80 px-3 text-xs font-semibold"
+        >
+          <span class="truncate">{statusLabel}</span>
+        </Select.Trigger>
+        <Select.Content class="border-stroke bg-white">
+          {#each statusOptions as option (option.value)}
+            <Select.Item value={option.value} label={option.label} />
+          {/each}
+        </Select.Content>
+      </Select.Root>
+    </div>
+  </div>
+</section>
+
+<section
+  class="border-stroke shadow-card mt-3 rounded-2xl border bg-white/90 p-3"
+>
+  <div class="flex flex-wrap items-end justify-between gap-2">
+    <div class="text-ink text-sm font-semibold">Players</div>
+    {#if data.activeRange?.note && rangeValue.startsWith("season")}
+      <div
+        class="text-muted-foreground text-[0.6rem] font-semibold tracking-[0.2em] uppercase"
+      >
+        {data.activeRange.note}
+      </div>
+    {/if}
+  </div>
+
+  <div class="mt-3 overflow-x-auto overflow-y-visible">
+    <table class="w-full min-w-[720px] text-xs">
+      <thead class="bg-white/70 text-left">
+        <tr
+          class="text-muted-foreground text-[0.6rem] font-semibold tracking-[0.2em] uppercase"
+        >
+          <th class="px-2 py-2">#</th>
+          <th class="px-2 py-2">Player</th>
+          <th class="px-2 py-2 text-right">Rating</th>
+          <th class="px-2 py-2 text-right">Win%</th>
+          <th class="px-2 py-2 text-right">Games</th>
+          <th class="px-2 py-2 text-right">Points</th>
+          <th class="px-2 py-2 text-right">Diff</th>
+          <th class="px-2 py-2 text-right">Form</th>
         </tr>
-      {/each}
-    </tbody>
-  </table>
-</div>
+      </thead>
+      <tbody>
+        {#if data.results.length === 0}
+          <tr>
+            <td class="text-muted-foreground px-2 py-4" colspan="8">
+              No players found.
+            </td>
+          </tr>
+        {:else}
+          {#each data.results as player, index (player.id)}
+            <tr
+              class={twMerge(
+                "border-stroke/50 border-b transition last:border-b-0 hover:bg-white/70",
+                !player.isActive && "opacity-60",
+              )}
+            >
+              <td class="px-2 py-2 font-semibold tabular-nums">
+                {index + 1}
+              </td>
+              <td class="px-2 py-2">
+                <a
+                  class="font-semibold"
+                  href={resolve(
+                    toPathname(
+                      `/chat/${slug}/user/${player.id}${buildQuery(
+                        rangeValue,
+                        statusValue,
+                      )}`,
+                    ),
+                  )}
+                >
+                  {player.name}
+                </a>
+              </td>
+              <td class="px-2 py-2 pr-8 text-right">
+                <span
+                  class="relative inline-flex items-start font-semibold tabular-nums"
+                >
+                  {player.rating}
+                  {#if player.playedLastDay}
+                    <span
+                      class={twMerge(
+                        "absolute top-0.5 left-full ml-1 text-[0.5rem] leading-none font-semibold tabular-nums",
+                        player.ratingChange > 0 && "text-green-600",
+                        player.ratingChange < 0 && "text-red-500",
+                      )}
+                    >
+                      {player.ratingChange > 0 ? "+" : ""}{player.ratingChange}
+                    </span>
+                  {/if}
+                </span>
+              </td>
+              <td class="px-2 py-2 text-right tabular-nums">
+                {player.winrate}%
+              </td>
+              <td class="px-2 py-2 text-right tabular-nums">
+                {player.games}
+              </td>
+              <td class="px-2 py-2 text-right tabular-nums">
+                {player.pointsFor} / {player.pointsAgainst}
+              </td>
+              <td class="px-2 py-2 text-right font-semibold tabular-nums">
+                {player.pointDiff > 0 ? "+" : ""}{player.pointDiff}
+              </td>
+              <td class="overflow-visible px-2 py-2 text-right">
+                <div class="flex justify-end gap-1">
+                  {#each player.recentForm as mark, markIndex (markIndex)}
+                    {@const matchDetail = player.recentMatches?.[markIndex]}
+                    <button
+                      type="button"
+                      class={twMerge(
+                        "block h-2 w-2 rounded-full p-0",
+                        mark === "W" ? "bg-green-500" : "bg-red-500",
+                      )}
+                      aria-label="Match details"
+                      onmouseenter={(event) =>
+                        matchDetail && showMatchTooltip(event, matchDetail)}
+                      onmouseleave={hideMatchTooltip}
+                      onfocus={(event) =>
+                        matchDetail && showMatchTooltip(event, matchDetail)}
+                      onblur={hideMatchTooltip}
+                    ></button>
+                  {/each}
+                </div>
+              </td>
+            </tr>
+          {/each}
+        {/if}
+      </tbody>
+    </table>
+  </div>
+</section>
+
+{#if matchTooltip}
+  <div
+    class="border-stroke text-ink pointer-events-none fixed z-[2147483647] w-48 rounded-xl border bg-white p-1.5 text-[0.6rem] leading-tight shadow-xl"
+    style={`left: ${matchTooltip.x}px; top: ${matchTooltip.y}px;`}
+    role="tooltip"
+  >
+    <div class="text-[0.5rem] font-semibold tracking-[0.2em] uppercase">
+      {matchTooltip.detail.result === "W" ? "Win" : "Loss"} ·
+      {matchTooltip.detail.score}
+    </div>
+    <div class="mt-0.5 font-semibold">
+      With {matchTooltip.detail.teammates.join(" + ")}
+    </div>
+    <div class="text-ink/70">
+      vs {matchTooltip.detail.opponents.join(" + ")}
+    </div>
+  </div>
+{/if}
